requirements.txt
gitpython
requests
pika